requirements.txt
cffi==1.15.0
cryptography==36.0.0
docopt==0.6.2
mercurial==5.9.3
numpy==1.21.3
pycparser==2.21
PyJWT==2.1.0
python-dotenv==0.19.2
six==1.16.0
tabulate==0.8.9
